Completed week 2

+++ ionic/conFusion/www/js/controllers.js
@@ -259,10 +259,14 @@ angular.module('conFusion.controllers', [])
 
 }])
 
-.controller('FavoritesController', ['$scope', 'menuFactory', 'favoriteFactory', 'baseURL', '$ionicListDelegate', function($scope, menuFactory, favoriteFactory, baseURL, $ionicListDelegate) {
+.controller('FavoritesController', ['$scope', 'menuFactory', 'favoriteFactory', 'baseURL', '$ionicListDelegate', '$ionicPopup', '$ionicLoading', '$timeout', function($scope, menuFactory, favoriteFactory, baseURL, $ionicListDelegate, $ionicPopup, $ionicLoading, $timeout) {
 
   $scope.baseURL = baseURL;
   $scope.shouldShowDelete = false;
+
+  $ionicLoading.show({
+    template: '<ion-spinner></ion-spinner> Loading...'
+  });
 
   $scope.favorites = favoriteFactory.getFavorites();
 
@@ -274,13 +278,31 @@ angular.module('conFusion.controllers', [])
   menuFactory.getDishes().query(
     function(response) {
       $scope.dishes = response;
+      $timeout(function(){
+        $ionicLoading.hide();
+      }, 1000);
     },
     function(response) {
-      $scope.message = "Error: " + response.status + " " + response.statusText;
+      $scope.message = "Error: " + response.status + " " + response.statusText;$scope.dishes = response;
+      $timeout(function(){
+        $ionicLoading.hide();
+      }, 1000);
     });
-    
+
   $scope.deleteFavorite = function(index) {
-    favoriteFactory.deleteFromFavorites(index);
+    var confirmPopup = $ionicPopup.confirm({
+      title: 'Confirm Delete',
+      template: 'Are you sure you want to delete this item?'
+    });
+
+    confirmPopup.then(function(res){
+      if (res){
+        console.log('OK to delete');
+        favoriteFactory.deleteFromFavorites(index);
+      } else {
+        console.log('Cancelled delete');
+      }
+    })
     $scope.shouldShowDelete = false;
   };
